Academy initial contribution, HTML tutorial added.
After adding the core academy page and the web development main page, the armies of the first contribution marched to the aid of those seeking to learn HTML through the medium of webpages.


However, the results for those desiring video format content, or PDF, Epub, and so on content had not yet been met.

Additionally, the many desires of humanity for succint and beautifully expressed knowledge had not yet been met in many fields. Not only did more lessons on web development await making, but lessons in many other fields too. 

Clearly, the God of Elkia needed allies, fellow Gods and kings for the upcoming tasks at hand...